// angular/orcamento/src/app/aw-components/util/types.ts
export type AwComponentStatus = 'primary' | 'secondary' | 'success' | 'danger' | 'warning';

export type AwComponentSize = 'xs' | 'sm' | 'md' | 'lg';

// TODO sera feito um processo de unificação de itens, e esse type será modificado
export type AwIcon =
  | 'compare'
  | 'review-docs'
  | 'print'
  | 'more-options'
  | 'create-doc'
  | 'copy'
  | 'cog-solid'
  | 'close'
  | 'archive-solid'
  | 'download'
  | 'exchange'
  | 'user-placeholder'
  | 'user'
  | 'star-empty'
  | 'star-full'
  | 'loop'
  | 'left-arrow'
  | 'spreadsheet'
  | 'curva-abc'
  | 'files-list'
  | 'dashed-eye'
  | 'reorder'
  | 'action'
  | 'add-provider'
  | 'attach'
  | 'eye-open'
  | 'history'
  | 'information'
  | 'open-budget'
  | 'resend'
  | 'send-ok'
  | 'send'
  | 'upload-document-check'
  | 'upload-document'
  | 'check'
  | 'filter'
  | 'timer'
  | 'filter-bars'
  | 'bell'
  | 'toggle-theme'
  | 'caret-down'
  | 'logo'
  | 'menu'
  | 'search'
  | 'edit'
  | 'calendar'
  | 'direction-arrow'
  | 'trash-alt'
  | 'plus'
  | 'drag-n-drop'
  | 'trash'
  | 'right-arrow'
  | 'budget-configuration'
  | 'budget-plan'
  | 'correct'
  | 'error'
  | 'caret-up'
  | 'plus-circle'
  | 'add'
  | 'sum'
  | 'notification'
  | 'spin'
  | 'file-pdf'
  | 'arrow-left2'
  | 'home'
  | 'icon-compare'
  | 'icon-review-docs'
  | 'icon-print'
  | 'icon-more-options'
  | 'icon-create-doc'
  | 'icon-copy'
  | 'icon-cog-solid'
  | 'icon-close'
  | 'icon-archive-solid'
  | 'icon-download'
  | 'icon-exchange'
  | 'icon-user-placeholder'
  | 'icon-user'
  | 'icon-star-empty'
  | 'icon-star-full'
  | 'icon-loop'
  | 'icon-left-arrow'
  | 'icon-spreadsheet'
  | 'icon-curva-abc'
  | 'icon-files-list'
  | 'icon-dashed-eye'
  | 'icon-reorder'
  | 'icon-action'
  | 'icon-add-provider'
  | 'icon-attach'
  | 'icon-eye-open'
  | 'icon-history'
  | 'icon-information'
  | 'icon-open-budget'
  | 'icon-resend'
  | 'icon-send-ok'
  | 'icon-send'
  | 'icon-upload-document-check'
  | 'icon-upload-document'
  | 'icon-check'
  | 'icon-filter'
  | 'icon-timer'
  | 'icon-filter-bars'
  | 'icon-bell'
  | 'icon-toggle-theme'
  | 'icon-caret-down'
  | 'icon-logo'
  | 'icon-menu'
  | 'icon-search'
  | 'icon-edit'
  | 'icon-calendar'
  | 'icon-direction-arrow'
  | 'icon-trash-alt'
  | 'icon-plus'
  | 'icon-drag-n-drop'
  | 'icon-trash'
  | 'icon-right-arrow'
  | 'icon-budget-configuration'
  | 'icon-budget-plan'
  | 'icon-correct'
  | 'icon-error'
  | 'icon-caret-up'
  | 'icon-plus-circle'
  | 'icon-add'
  | 'icon-sum'
  | 'icon-notification'
  | 'icon-spin'
  | 'icon-file-pdf'
  | 'icon-arrow-left2'
  | 'icon-home'
  | 'aw-icon-check-inactive'
  | 'aw-icon-checked'
  | 'aw-icon-icon-tooltip'
  | 'aw-icon-msg-error'
  | 'aw-icon-msg-success'
  | 'aw-icon-msg-black'
  | 'aw-icon-last-call-inative'
  | 'aw-icon-last-call'
  | 'aw-icon-frozen'
  | 'aw-icon-expand-arrow-left'
  | 'aw-icon-expand-arrow-left-hover'
  | 'aw-icon-expand-arrow-right'
  | 'aw-icon-expand-arrow-right-hover'
  | 'aw-icon-logo'
  | 'aw-icon-criar-doc'
  | 'aw-icon-imprimir'
  | 'aw-icon-abrir-budget'
  | 'aw-icon-revisar-budget'
  | 'aw-icon-engrenagem'
  | 'aw-icon-duplicate'
  | 'aw-icon-calendar-x'
  | 'aw-icon-calendar-check'
  | 'aw-icon-seta-cima-clara'
  | 'aw-icon-seta-baixo-clara'
  | 'aw-icon-seta-esquerda'
  | 'aw-icon-seta-direita'
  | 'aw-icon-circulo-vermelho'
  | 'aw-icon-alerta-cheio'
  | 'aw-icon-erro'
  | 'aw-icon-alerta'
  | 'aw-icon-sucesso'
  | 'aw-icon-retangulo-cinza'
  | 'aw-icon-retangulo-azul'
  | 'aw-icon-dolly-flatbed'
  | 'aw-icon-lock';

const iconPaths: [AwIcon, number][] = [
  ['aw-icon-last-call', 2],
  ['aw-icon-expand-arrow-left-hover', 3],
  ['aw-icon-expand-arrow-right-hover', 3],
  ['aw-icon-logo', 6],
  ['aw-icon-revisar-budget', 7],
  ['aw-icon-calendar-x', 2],
  ['aw-icon-calendar-check', 2],
  ['aw-icon-last-call-inative', 2],
  ['icon-send-ok', 5],
  ['icon-bell', 4],
];

export const iconsPathMap = new Map<AwIcon, string[]>(
  iconPaths.map(([icon, paths]) => [icon, Array.from({ length: paths }, (_, index) => `path${index + 1}`)])
);